==> notebooks/ebdt_data/sub_network/n_model/03-learning-parameters-exp1.ipynb <==
#### Prerequesites:
* Define schema (make figure) 
* Define structure (make figure)
* Define schema class in python (src/models/model_schemata.py)
* Convert schema + structure into model template (models/templates/p_model.pl)

#### Procedure:
* Load data
* Adjust model script for lfi accordingly
* Create evidence file (if I don't need different evidence files, move to builmodel notebook)
* LFI

Multiple instances, 1 pst, X enzymes, one sample per instance, multiple enzymes per instance, test how many enzymes can have in one instances, how computing time depends on N p::f for 60 samples

%load_ext autoreload
%autoreload 2

# Importing libraries
import pandas as pd
import numpy as np
import os
import random

from src.models import model_schemata as schema
from src.models import build_model as build
from src.visualization import visualize as viz

os.chdir('/Users/magdalena/OneDrive - Queen Mary, University of London/bezzlab/research/projects/phospho_pi/')
# os.chdir('/home/mhuebner/Desktop/bezzlab/research/projects/phospho_pi/')

#### Training/Testing

# reading csv from files into dict
data = {}
data['e_ksea'] = pd.read_csv('data/processed/ebdt_data/sub_network_n/e_ksea.csv')
data['p_fc'] = pd.read_csv('data/processed/ebdt_data/sub_network_n/p_fc.csv')
data['es_interaction'] = pd.read_csv('data/processed/ebdt_data/sub_network_n/es_interaction.csv')
data['p_regulates'] = pd.read_csv('data/processed/ebdt_data/sub_network_n/p_regulates.csv')
data['e_ksea']

data['p_fc']['phosphosite'].value_counts()

Splitting data into training/testing

# getting sample names (union of all samples in all datasets)
samples = list(set(data['e_ksea']['sample']).union(set(data['p_fc']['sample'])))
samples.sort()
# sample x% of samples randomly without replacement with seed
random.seed(612)
train = random.sample(samples, int(len(samples)*1))
print(train)
# filter data
training_data = {}
training_data['e_ksea'] = data['e_ksea'][data['e_ksea']['sample'].isin(train)].reset_index(drop=True)
training_data['p_fc'] = data['p_fc'][data['p_fc']['sample'].isin(train)].reset_index(drop=True)

# Mapping data to Problog predicates
predicates = {}
predicates['e_ksea'] = schema.EKseaPredicate()
predicates['e_ksea'].add_data(training_data['e_ksea'], 'enzyme', 'sample', 'value')
predicates['p_fc'] = schema.PFoldChangePredicate()
predicates['p_fc'].add_data(training_data['p_fc'], 'phosphosite', 'sample', 'value')

samples = list(set(training_data['e_ksea']['sample']).union(set(training_data['p_fc']['sample'])))
enzymes = list(set(data['p_regulates']['protein'].tolist() + data['es_interaction']['enzyme'].tolist()))
evidence_dict = {}
for s in samples:
    evidence_dict[s] = {}
    for e0 in enzymes:
        evidence_dict[s][e0] = {}
        phosphosites = data['p_regulates']['phosphosite'][data['p_regulates']['protein'] == e0].tolist()
        for p in phosphosites:
            evid_generator_e = build.ProblogStatementGenerator(predicates['e_ksea'])
            evidence_e0 = evid_generator_e.generate_facts(build.EvidenceTemplate, select=[s, e0])
            evid_generator_p = build.ProblogStatementGenerator(predicates['p_fc'])
            evidence_p = evid_generator_p.generate_facts(build.EvidenceTemplate, select=[s, p])
            enzymes_p = data['es_interaction']['enzyme'][data['es_interaction']['phosphosite'] == p].tolist()
            evidence_e = []
            for e in enzymes_p:
                evidence_e.extend(evid_generator_e.generate_facts(build.EvidenceTemplate, select=[s, e]))
            evidence_dict[s][e0][p] = evidence_e0 + evidence_p + evidence_e
        if len(phosphosites) == 0:
            evid_generator_e = build.ProblogStatementGenerator(predicates['e_ksea'])
            evidence_e0 = evid_generator_e.generate_facts(build.EvidenceTemplate, select=[s, e0])
            evidence_dict[s][e0][e0] = evidence_e0

# write evidence_dict to text file: loop over samples and separate by line of hyphens
with open('models/ebdt_data/sub_network/n_model/evidence_exp1.pl', 'w') as f:
    for s in evidence_dict:
        for e in evidence_dict[s]:
            for p in evidence_dict[s][e]:
                for fact in evidence_dict[s][e][p]:
                    f.write(fact + '\n')
                f.write('--------------------' + '\n')

# write evidence_dict to text file: loop over samples and separate by line of hyphens
with open('models/ebdt_data/sub_network/n_model/evidence_exp1_training.pl', 'w') as f:
    for s in evidence_dict:
        for e in evidence_dict[s]:
            for p in evidence_dict[s][e]:
                for fact in evidence_dict[s][e][p]:
                    f.write(fact + '\n')
                f.write('--------------------' + '\n')

LFI

DIR = 'models/ebdt_data/sub_network/n_model/'
max_iter = 1
# learning from interpretation
cmd = f'problog lfi {DIR}n_model_lfi_exp1.pl {DIR}evidence_exp1.pl -O {DIR}n_model_exp1_1i.pl --logger {DIR}log_exp1-1i.txt -k ddnnf -v -n {max_iter}'
os.system(cmd)

DIR = 'models/ebdt_data/sub_network/n_model/'
# learning from interpretation
cmd = f'problog ground {DIR}n_model_lfi_exp1.pl -o {DIR}n_model_exp1_ground.pl --format pl'
os.system(cmd)
