==> notebooks/ebdt_data/sub_network/n_model/05-predictions-exp1-mpe.ipynb <==
#### Prerequesites:
* Define schema (make figure) 
* Define structure (make figure)
* Define schema class in python (src/models/model_schemata.py)
* Convert schema + structure into model template (models/templates/p_model.pl)

#### Procedure:
* Load data
* Adjust model script for lfi accordingly
* Create evidence file (if I don't need different evidence files, move to builmodel notebook)
* LFI

Multiple instances, 1 pst, X enzymes, one sample per instance, multiple enzymes per instance, test how many enzymes can have in one instances, how computing time depends on N p::f for 60 samples

%load_ext autoreload
%autoreload 2

# Importing libraries
import pandas as pd
import numpy as np
import os
import pickle

from src.models import model_schemata as schema
from src.models import build_model as build
from src.models import parameter_learning as lfi
from src.models import inference as inf
from src.visualization import visualize as viz


os.chdir('/Users/magdalena/OneDrive - Queen Mary, University of London/bezzlab/research/projects/phospho_pi/')
# os.chdir('/home/mhuebner/Desktop/bezzlab/research/projects/phospho_pi/')

#### MPE (I)

# reading csv from files into dict
data = {}
data['e_ksea'] = pd.read_csv('data/processed/ebdt_data/sub_network_n/e_ksea.csv')
data['p_fc'] = pd.read_csv('data/processed/ebdt_data/sub_network_n/p_fc.csv')

# Adding evidence
model_inference = 'models/ebdt_data/sub_network/n_model/n_model_exp1_mpe.pl'

Adding queries

sample_ids = data['e_ksea']['sample'].unique().tolist()
enzymes = data['e_ksea']['enzyme'].unique().tolist()
queries = inf.generate_queries(schema.EActivityPredicate, sample_ids=sample_ids, enzymes=enzymes)
len(queries)

# make evidence_dict['AZD5438'] into list
query_list = []
for s in sample_ids:
    # return where queries contains sample
    matches = [s in q for q in queries]
    # select queries where sample is AC220
    q = [item for item, condition in zip(queries, matches) if condition]
    query_list.append(q)
query_list[:2]

Adding evidence (all phosphosite and relevant enzymes)

# Mapping data to Problog predicates
predicates = {}
predicates['p_fc'] = schema.PFoldChangePredicate()
predicates['p_fc'].add_data(data['p_fc'], 'phosphosite', 'sample', 'value')
predicates['e_ksea'] = schema.EKseaPredicate()
predicates['e_ksea'].add_data(data['e_ksea'], 'enzyme', 'sample', 'value')

evidence_dict = {}
for s in sample_ids:
    evidence_dict[s] = {}
    #evid_generator = build.ProblogStatementGenerator(predicates['e_ksea'])
    #evidence_e = evid_generator.generate_facts(build.EvidenceTemplate, select=[s])
    evid_generator = build.ProblogStatementGenerator(predicates['p_fc'])
    evidence_p = evid_generator.generate_facts(build.EvidenceTemplate, select=[s])
    # evidence_dict[s] = evidence_e + evidence_p
    evidence_dict[s] = evidence_p

# make evidence_dict['AZD5438'] into list
evidence_list = []
for s in evidence_dict.keys():
    evidence_list.append(evidence_dict[s])
evidence_list[:2]

query = [query_list[0][0]]
query

inference_output = 'models/ebdt_data/sub_network/n_model/inference_log.txt'
inf_data_dict = inf.infer_mpe(model_inference, query, inference_output, evidence=[])
pickle.dump(inf_data_dict, open('models/ebdt_data/sub_network/n_model/predictions_mpe_allpev_exp5.pkl', 'wb'))

inf_data_dict['e_activity']
